Update `mithril-client-cli` directory after previous modifications
- remove services
- move `http_client` to /src
- remove `aggregator-client` module
- update `builder`: remove unused dependencies
