--- mithril-client-cli/src/aggregator_client/mod.rs
@@ -1,10 +1,0 @@
-//! Transport layers to the Aggregator nodes
-mod certificate_client;
-mod http_client;
-mod mithril_stake_distribution_client;
-mod snapshot_client;
-
-pub use certificate_client::*;
-pub use http_client::*;
-pub use mithril_stake_distribution_client::*;
-pub use snapshot_client::*;
