+++ mithril-client-cli/src/dependencies/builder.rs
@@ -1,27 +1,11 @@
-use std::{collections::HashMap, sync::Arc};
-
-use anyhow::Context;
 use serde::Deserialize;
 use slog::Logger;
-
-use mithril_common::{
-    api_version::APIVersionProvider,
-    certificate_chain::{CertificateVerifier, MithrilCertificateVerifier},
-    digesters::{CardanoImmutableDigester, ImmutableDigester},
-    StdResult,
-};
+use std::{collections::HashMap, sync::Arc};
 use thiserror::Error;
 
-use crate::{
-    aggregator_client::{
-        AggregatorClient, AggregatorHTTPClient, CertificateClient, MithrilStakeDistributionClient,
-        SnapshotClient,
-    },
-    services::{
-        AppMithrilStakeDistributionService, MithrilClientSnapshotService,
-        MithrilStakeDistributionService, SnapshotService,
-    },
-};
+use mithril_common::{api_version::APIVersionProvider, StdResult};
+
+use crate::http_client::{AggregatorClient, AggregatorHTTPClient};
 
 /// Configuration error
 #[derive(Debug, Error)]
@@ -87,27 +71,6 @@ pub struct DependenciesBuilder {
 
     /// HTTP Aggregator client
     pub aggregator_client: Option<Arc<dyn AggregatorClient>>,
-
-    /// SnapshotClient
-    pub snapshot_client: Option<Arc<SnapshotClient>>,
-
-    /// MithrilStakeDistributionClient
-    pub mithril_stake_distribution_client: Option<Arc<MithrilStakeDistributionClient>>,
-
-    /// CertificateClient
-    pub certificate_client: Option<Arc<CertificateClient>>,
-
-    /// CertificateVerifier
-    pub certificate_verifier: Option<Arc<dyn CertificateVerifier>>,
-
-    /// ImmutableDigester
-    pub immutable_digester: Option<Arc<dyn ImmutableDigester>>,
-
-    /// [SnapshotService]
-    pub snapshot_service: Option<Arc<dyn SnapshotService>>,
-
-    /// [MithrilStakeDistributionService]
-    pub mithril_stake_distribution_service: Option<Arc<dyn MithrilStakeDistributionService>>,
 }
 
 impl DependenciesBuilder {
@@ -116,13 +79,6 @@ impl DependenciesBuilder {
         Self {
             config,
             aggregator_client: None,
-            snapshot_client: None,
-            mithril_stake_distribution_client: None,
-            certificate_client: None,
-            certificate_verifier: None,
-            immutable_digester: None,
-            snapshot_service: None,
-            mithril_stake_distribution_service: None,
         }
     }
 
@@ -152,171 +108,6 @@ impl DependenciesBuilder {
         }
 
         Ok(self.aggregator_client.as_ref().cloned().unwrap())
-    }
-
-    async fn build_snapshot_client(&mut self) -> StdResult<Arc<SnapshotClient>> {
-        let client = SnapshotClient::new(self.get_aggregator_client().await?);
-
-        Ok(Arc::new(client))
-    }
-
-    /// Get a clone of the [SnapshotClient] dependency
-    pub async fn get_snapshot_client(&mut self) -> StdResult<Arc<SnapshotClient>> {
-        if self.snapshot_client.is_none() {
-            self.snapshot_client = Some(
-                self.build_snapshot_client()
-                    .await
-                    .with_context(|| "Dependencies Builder can not build Snapshot Client")?,
-            );
-        }
-
-        Ok(self.snapshot_client.as_ref().cloned().unwrap())
-    }
-
-    async fn build_mithril_stake_distribution_client(
-        &mut self,
-    ) -> StdResult<Arc<MithrilStakeDistributionClient>> {
-        let client = MithrilStakeDistributionClient::new(self.get_aggregator_client().await?);
-
-        Ok(Arc::new(client))
-    }
-
-    /// Get a clone of the [SnapshotClient] dependency
-    pub async fn get_mithril_stake_distribution_client(
-        &mut self,
-    ) -> StdResult<Arc<MithrilStakeDistributionClient>> {
-        if self.mithril_stake_distribution_client.is_none() {
-            self.mithril_stake_distribution_client = Some(
-                self.build_mithril_stake_distribution_client()
-                    .await
-                    .with_context(|| {
-                        "Dependencies Builder can not build Mithril Stake distribution Client"
-                    })?,
-            );
-        }
-
-        Ok(self
-            .mithril_stake_distribution_client
-            .as_ref()
-            .cloned()
-            .unwrap())
-    }
-
-    async fn build_certificate_client(&mut self) -> StdResult<Arc<CertificateClient>> {
-        let client = CertificateClient::new(self.get_aggregator_client().await?);
-
-        Ok(Arc::new(client))
-    }
-
-    /// Get a clone of the [CertificateClient] dependency
-    pub async fn get_certificate_client(&mut self) -> StdResult<Arc<CertificateClient>> {
-        if self.certificate_client.is_none() {
-            self.certificate_client = Some(
-                self.build_certificate_client()
-                    .await
-                    .with_context(|| "Dependencies Builder can not build Certificate Client")?,
-            );
-        }
-
-        Ok(self.certificate_client.as_ref().cloned().unwrap())
-    }
-
-    async fn build_certificate_verifier(&mut self) -> StdResult<Arc<dyn CertificateVerifier>> {
-        let verifier = MithrilCertificateVerifier::new(
-            self.get_logger().await?,
-            self.get_certificate_client().await?,
-        );
-
-        Ok(Arc::new(verifier))
-    }
-
-    /// Get a clone of the [CertificateVerifier] dependency
-    pub async fn get_certificate_verifier(&mut self) -> StdResult<Arc<dyn CertificateVerifier>> {
-        if self.certificate_verifier.is_none() {
-            self.certificate_verifier = Some(
-                self.build_certificate_verifier()
-                    .await
-                    .with_context(|| "Dependencies Builder can not build Certificate Verifier")?,
-            );
-        }
-
-        Ok(self.certificate_verifier.as_ref().cloned().unwrap())
-    }
-
-    async fn build_immutable_digester(&mut self) -> StdResult<Arc<dyn ImmutableDigester>> {
-        let digester = CardanoImmutableDigester::new(None, self.get_logger().await?);
-
-        Ok(Arc::new(digester))
-    }
-
-    /// Get a clone of the [ImmutableDigester] dependency
-    pub async fn get_immutable_digester(&mut self) -> StdResult<Arc<dyn ImmutableDigester>> {
-        if self.immutable_digester.is_none() {
-            self.immutable_digester = Some(
-                self.build_immutable_digester()
-                    .await
-                    .with_context(|| "Dependencies Builder can not build Immutable Digester")?,
-            );
-        }
-
-        Ok(self.immutable_digester.as_ref().cloned().unwrap())
-    }
-
-    async fn build_snapshot_service(&mut self) -> StdResult<Arc<dyn SnapshotService>> {
-        let snapshot_service = MithrilClientSnapshotService::new(
-            self.get_snapshot_client().await?,
-            self.get_certificate_client().await?,
-            self.get_certificate_verifier().await?,
-            self.get_immutable_digester().await?,
-        );
-
-        Ok(Arc::new(snapshot_service))
-    }
-
-    /// Get a clone of the [SnapshotService] dependency
-    pub async fn get_snapshot_service(&mut self) -> StdResult<Arc<dyn SnapshotService>> {
-        if self.snapshot_service.is_none() {
-            self.snapshot_service = Some(
-                self.build_snapshot_service()
-                    .await
-                    .with_context(|| "Dependencies Builder can not build Snapshot Service")?,
-            );
-        }
-
-        Ok(self.snapshot_service.as_ref().cloned().unwrap())
-    }
-
-    async fn build_mithril_stake_distribution_service(
-        &mut self,
-    ) -> StdResult<Arc<dyn MithrilStakeDistributionService>> {
-        let service = AppMithrilStakeDistributionService::new(
-            self.get_mithril_stake_distribution_client().await?,
-            self.get_certificate_client().await?,
-            self.get_certificate_verifier().await?,
-        );
-
-        Ok(Arc::new(service))
-    }
-
-    /// Get a clone of the [MithrilStakeDistributionService] dependency
-    pub async fn get_mithril_stake_distribution_service(
-        &mut self,
-    ) -> StdResult<Arc<dyn MithrilStakeDistributionService>> {
-        if self.mithril_stake_distribution_service.is_none() {
-            self.mithril_stake_distribution_service = Some(
-                self.build_mithril_stake_distribution_service()
-                    .await
-                    .with_context(|| {
-                        "Dependencies Builder can not build Mithril Stake Distribution Service"
-                    })?,
-            );
-        }
-
-        Ok(self
-            .mithril_stake_distribution_service
-            .as_ref()
-            .cloned()
-            .unwrap())
     }
 }
