--- mithril-client-cli/src/services/mod.rs
@@ -1,13 +1,0 @@
-//! Services
-//! This module contains the different services tied with their bounded context.
-mod mithril_stake_distribution;
-mod snapshot;
-
-#[cfg(test)]
-mod mock;
-
-pub use mithril_stake_distribution::*;
-pub use snapshot::*;
-
-#[cfg(test)]
-pub use mock::*;
