+++ mithril-client-cli/src/utils/mod.rs
@@ -3,10 +3,8 @@
 
 mod progress_reporter;
 mod snapshot;
-mod stream_reader;
 mod unpacker;
 
 pub use progress_reporter::*;
 pub use snapshot::*;
-pub use stream_reader::*;
 pub use unpacker::*;
